Crate narrow

narrow

An implementation of Apache Arrow.

Structs

Bitmap

An immutable collection of bits.

BitmapIter

BooleanArray

Array with boolean values.

Buffer

A contiguous immutable memory buffer for data.

DenseUnionArray

Dense union array for enum type T with N variants.

DenseUnionIter

Iterator over elements of a dense union array.

DictionaryArray

Array where values are encoded using a dictionary.

DictionaryArrayIter

Iterator over elements of a dictionary encoded array.

FixedSizeListArray

Array with fixed-size lists of other array types.

FixedSizeListArrayIter

Iterator over elements of an array with fixed-sized lists of other arrays.

FixedSizePrimitiveArray

Array with primitive values.

NullArray

A sequence of nulls.

Nullable

Wrapper for nullable data.

Offset

Wrapper for offset values of variable sized arrays.

OffsetIter

Iterator over offsets values of variable sized arrays.

SparseUnionArray

Sparse union array for enum type T.

SparseUnionIter

Iterator over elements of a sparse union array.

StringArray

Array with variable sized string (UTF-8) data.

StructArray

Array with structs that have fields of other array types.

UnionArray

Array with an ordered sequence of variants of T. D encodes the union array type.

Validity

Validity wrapper for nullable and non-nullable data.

VariableSizeBinaryArray

Array with variable-sized binary data.

VariableSizeBinaryArrayIter

Iterator over elements of an array with variable-sized binary data.

VariableSizeListArray

Array with variable-sized lists of other arrays.

VariableSizeListArrayIter

Iterator over elements of an array with variable-sized lists of other arrays.

Traits

Array

A sequence of values with known length all having the same type.

ArrayData

Types storing array data.

ArrayIndex

Index trait to get owned values of an array.

ArrayType

Types for which sequences of values can be stored in arrays.

DictionaryIndexValue

Types representing dictionary index values.

NestedArray

Types storing nested sequences of values.

OffsetValue

Types representing offset values.

Primitive

Subtrait for primitive types.

StructArrayType

Struct types that can be stored in arrays.

UnionArrayIndex

Index trait for union arrays.

UnionArrayType

Union types that can be stored in arrays.

Type Definitions

BinaryArray

Array with variable sized binary data. Uses i32 offsets.

Float32Array

Array with f32 values.

Float64Array

Array with f64 values.

Int8Array

Array with i8 values.

Int16Array

Array with i16 values.

Int32Array

Array with i32 values.

Int64Array

Array with i64 values.

LargeBinaryArray

Array with variable sized binary data. Uses i64 offsets.

LargeListArray

Array with variable-sized lists of other array types. Uses i64 offsets.

LargeUtf8Array

Array with UTF-8 strings. Uses i64 offsets.

ListArray

Array with variable-sized lists of other array types. Uses i32 offsets.

Uint8Array

Array with u8 values.

Uint16Array

Array with u16 values.

Uint32Array

Array with u32 values.

Uint64Array

Array with u64 values.

Utf8Array

Array with UTF-8 strings. Uses i32 offsets.

Derive Macros

Array

Derive macro for the Array trait.