Trait narrow::bitmap::iter::BitPackedExt

pub trait BitPackedExt<T> where
    Self: Iterator<Item = T>,
    T: Borrow<bool>, {
    fn bit_packed(self) -> BitPacked<Self, T>
    where
        Self: Sized,
    { ... }
}

An Iterator extension trait for BitPacked.

Provided Methods

fn bit_packed(self) -> BitPacked<Self, T> where
Self: Sized,

Packs the items in this iterator that can be borrowed as boolean values as bits in bytes using least-significant bit (LSB) numbering.