Struct narrow::nullable::Nullable

pub struct Nullable<T, U> where
    U: Buffer<u8>,  { /* private fields */ }

Wrapper for nullable data.

Store data with a validity Bitmap that uses a single bit per value in T that indicates the nullness or non-nullness of that value.