Trait narrow::ArrayData

pub trait ArrayData {
    fn len(&self) -> usize;

    fn is_valid(&self, index: usize) -> bool;

    fn valid_count(&self) -> usize;

    fn is_empty(&self) -> bool { ... }

    fn is_null(&self, index: usize) -> bool { ... }

    fn null_count(&self) -> usize { ... }
}

Types storing array data.

Required methods

fn len(&self) -> usize

Returns the number of elements.

fn is_valid(&self, index: usize) -> bool

fn valid_count(&self) -> usize

Returns the number of valid elements.

Provided methods

fn is_empty(&self) -> bool

Returns true if the number of elements is 0.

fn is_null(&self, index: usize) -> bool

Returns true if the element at position index is null.

fn null_count(&self) -> usize

Returns the number of null elements.