Trait narrow::Null

pub trait Null: Length {
    unsafe fn is_valid_unchecked(&self, index: usize) -> bool;

    fn is_null(&self, index: usize) -> Option<bool> { ... }
    unsafe fn is_null_unchecked(&self, index: usize) -> bool { ... }
    fn null_count(&self) -> usize { ... }
    fn is_valid(&self, index: usize) -> Option<bool> { ... }
    fn valid_count(&self) -> usize { ... }
    fn any_null(&self) -> bool { ... }
    fn all_null(&self) -> bool { ... }
    fn any_valid(&self) -> bool { ... }
    fn all_valid(&self) -> bool { ... }
}

Null-ness of elements in a collection.

Required methods

unsafe fn is_valid_unchecked(&self, index: usize) -> bool

Returns true if the element at position index is valid, without performing any bounds checking.

Safety

The index must be in bounds.

Calling this method with an out-of-bounds index is undefined behavior.

Provided methods

fn is_null(&self, index: usize) -> Option<bool>

Returns true if the element at position index is null.

unsafe fn is_null_unchecked(&self, index: usize) -> bool

Returns true if the element at position index is null, without performing any bounds checking.

Safety

The index must be in bounds.

Calling this method with an out-of-bounds index is undefined behavior.

fn null_count(&self) -> usize

Returns the number of null elements.

fn is_valid(&self, index: usize) -> Option<bool>

Returns true if the element at position index is valid.

fn valid_count(&self) -> usize

Returns the number of valid elements.

fn any_null(&self) -> bool

Returns true if the array contains at least one null element.

fn all_null(&self) -> bool

Returns true if all the elements are null.

fn any_valid(&self) -> bool

Returns true if the array contains at least one valid element.

fn all_valid(&self) -> bool

Returns true if all the elements are valid.